LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/setCustomValidity-normalize-newlines-expected.txt



FAIL setCustomValidity should normalize newlines from the given error message assert_equals: expected "First line\nSecond line\nThird line\nFourth line" but got "First line\rSecond line\r\nThird line\nFourth line"
